Volume feature + 42 header feature
Now we can access to a work directory and update codebase or whatever files. The updates are also made into the container (if available directory provided).

Trying to install a node environment but with the node docker image pulled directly from Docker Hub, it erases all tools installed before, so I think I have to manually install it.

Adding alse the 42 headers, but need to improve the script (what if the login is given empty?). Need also to update if the folder is given with a / final (have to remove it), get back the last part of the folder for putting it into the container.